Fixes to line alignment

+++ rm_lines/headers/advanced/line_info.h
@@ -4,6 +4,6 @@
 struct LineInfo {
     Line line;
     CrdtId groupId;
-    int offsetX;
-    int offsetY;
+    float offsetX;
+    float offsetY;
 };
